+++ bundles/framework/feedbackService/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('feedbackService', () => Oskari.clazz.create('Oskari.mapframework.bundle.feedbackService.FeedbackServiceBundleInstance'));